backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.3.3
pydantic==2.11.7
pydantic-settings==2.2.1
python-dotenv==1.0.0
pdfplumber==0.10.3
pymilvus==2.3.5
torch==2.11.0
sentence-transformers==3.0.1
google-generativeai==0.8.4
prometheus-client==0.19.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
isort==5.13.2
